app_users/forms.py
# from django import forms
from django.contrib.auth.forms import UserCreationForm, UserChangeForm
from .models import User
from unfold.admin import forms

class CustomUserCreationForm(UserCreationForm):
    class Meta:
        model = User
        fields = ('username', 'email', 'first_name', 'last_name')

        

class CustomUserChangeForm(UserChangeForm):
    class Meta:
        model = User
        fields = ('username', 'email', 'first_name', 'last_name', 'is_active')
